requirements.txt
flask==2.3.3
pyyaml==6.0.1
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==2.0.7
python-dotenv==1.0.0